requirements.txt
openai==0.28
pillow
transformers
accelerate